Issue #7 Change 'AttestationConveyancePreference' type
Constants are not addressable, i.e. the '&' operator cannot be used to take
the address of a const. This makes it impossible to create an instance of
a Request with a value other than its zero value. The commit changes the
property type in the request struct to a string pointer (still allowing for
nil values because it is an optional value for API requests).

Also adds a convenience method for getting a string
(pointer) for the preference constant, so that at least some 'enum-like'
behaviour is retained.